first/my-voice-journal/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.10
psycopg2==2.9.6
pydantic==1.10.7
python-multipart==0.0.6
google-cloud-speech==2.16.1
spacy==3.5.1
